Feature/backend auth (#7)
* added refresh token and jwt access token

For now interaction with the database is not done, scraping the surface

* additional methods

still no user table because of being afraid of merge conflict doing damage on the dbsnapshot

* added policy based access and an endpoint to test it

* changes to the refresh token

refresh token changed to JWT

* auth is working with frontend, products can now be edited, deleted and added from the frontend

* addOn Crud and auth endpoints done

* Some minor changes + frontend and policy based access on routes

* some fixes in signalR and frontend

+++ pizzeria-backend/Controllers/ProductsController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using pizzeria_backend.Models;
 using pizzeria_backend.Models.Interfaces;
 using pizzeria_backend.Services;
@@ -7,13 +8,20 @@ namespace pizzeria_backend.Controllers
 {
     [Route("api/products/v1.0")]
     [ApiController]
-    public class ProductsController(IProductService productService, IAzureBlobStorageService azureBlobStorageService) : ControllerBase
+    public class ProductsController : ControllerBase
     {
-        private readonly IProductService _productService = productService;
-        private readonly IAzureBlobStorageService _azureBlobStorageService = azureBlobStorageService;
+        private readonly IProductService _productService;
+        private readonly IAzureBlobStorageService _azureBlobStorageService;
 
+        public ProductsController(IProductService productsService, IAzureBlobStorageService azureBlobStorageService)
+        {
+
+            _productService = productsService;
+            _azureBlobStorageService = azureBlobStorageService;
+        }
 
         [HttpPost()]
+        [Authorize(Policy = "Admin")]
         [Produces("application/json")]
         public async Task<IActionResult> AddProductAsync([FromForm] ProductDto? product)
         {
@@ -21,7 +29,9 @@ namespace pizzeria_backend.Controllers
             {
                 return BadRequest("No body provided");
             }
-            var imageLink = (await _azureBlobStorageService.UploadBlobAsync(product.Image.OpenReadStream(), product.Image.FileName + Guid.NewGuid().ToString())).ToString();
+
+            var imageLink = (await _azureBlobStorageService.UploadBlobAsync(product.Image.OpenReadStream(), product.Image.FileName)).ToString();
+            Console.WriteLine(imageLink);
             var pr = await _productService.AddProductAsync(ConvertToProduct(product, imageLink));
 
             return Ok(pr);
@@ -33,11 +43,20 @@ namespace pizzeria_backend.Controllers
         public async Task<IActionResult> GetProduct(int id)
         {
             var pr = await _productService.GetProduct(id);
+
             if (pr == null)
             {
                 return NotFound("Product not found");
             }
 
+            return Ok(pr);
+        }
+
+        [HttpGet("")]
+        [Produces("application/json")]
+        public async Task<IActionResult> GetAllProduct()
+        {
+            var pr = await _productService.GetAllProducts();
 
             return Ok(pr);
         }
@@ -47,26 +66,34 @@ namespace pizzeria_backend.Controllers
         public async Task<IActionResult> GetMenu()
         {
             var menu = await _productService.GetMenu();
+
             if (menu == null)
             {
                 return NotFound("There are no items in the menu");
             }
+
             return Ok(menu);
         }
 
         [HttpDelete("{id}")]
+        [Authorize(Policy = "Admin")]
         [Produces("application/json")]
         public async Task<IActionResult> DeleteProduct(int id)
         {
             var pr = await _productService.DeleteProduct(id);
+
             if (pr == null)
             {
                 return NotFound("Product not found");
             }
+
+            await _azureBlobStorageService.DeleteBlobAsync(pr.Image.Split("/").Last());
+
             return Ok(pr);
         }
 
         [HttpPatch("{id}")]
+        [Authorize(Policy = "Admin")]
         [Produces("application/json")]
         public async Task<IActionResult> UpdateProduct(int id, [FromForm] ProductDto product)
         {
@@ -76,10 +103,19 @@ namespace pizzeria_backend.Controllers
                 return NotFound("Product not found");
             }
 
+            if (product.Image != null)
+            {
+                await _azureBlobStorageService.DeleteBlobAsync(pr.Image.Split("/").Last());
+                var image = (await _azureBlobStorageService.UploadBlobAsync(product.Image.OpenReadStream(), product.Image.FileName)).ToString();
+                pr = ConvertToProduct(product, image, id);
 
-            await _azureBlobStorageService.DeleteBlobAsync(pr.Image.Split("/").Last());
-            var image = (await _azureBlobStorageService.UploadBlobAsync(product.Image.OpenReadStream(), product.Image.FileName)).ToString();
-            pr = ConvertToProduct(product, image, id);
+            }
+            else
+            {
+                pr = ConvertToProduct(product, pr.Image, id);
+
+            }
+
             pr = await _productService.UpdateProduct(pr);
 
             return Ok(pr);
@@ -99,9 +135,9 @@ namespace pizzeria_backend.Controllers
             };
         }
 
-
         private static Product ConvertToProduct(ProductDto productDto, string image, int id)
         {
+
             return new Product
             {
                 Id = id,
@@ -114,5 +150,6 @@ namespace pizzeria_backend.Controllers
                 IsInMenu = productDto.IsInMenu,
             };
         }
+
     }
 }
